requirements.txt
certifi==2018.4.16
dj-database-url==0.5.0
Django==2.0.7
django-heroku==0.3.1
factory-boy==2.11.1
Faker==0.8.16
gunicorn==19.9.0
pipenv==2018.7.1
psycopg2==2.7.5
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
text-unidecode==1.2
uuid==1.30
virtualenv==16.0.0
virtualenv-clone==0.3.0
whitenoise==3.3.1
